requirements.txt
tensorflow-gpu==1.14.0
keras==2.2.4
tqdm
sklearn
scikit-image
simpleitk

torch==1.3.1
torchsummary
scipy
matplotlib